Move reservation related utils to correct file
Refs #397.
